Refactor app to use router, Refactor file strucutre
