+++ src/Components/URLParams/URLParams.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import { useParams } from 'react-router-dom';
+import SingleMovie from '../SingleMovie/SingleMovie';
+
+const URLParams = ({component}) => {
+  const id = useParams().id;
+  return (
+    <div>
+      <SingleMovie id={id}/>
+    </div>
+  )
+}
+
+export default URLParams
